% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_class}
\alias{add_class}
\title{Adds an S3 class to an object}
\usage{
add_class(x, class)
}
\arguments{
\item{x}{An object to which the class should be added.}

\item{class}{A character vector of class names to be added.}
}
\value{
The object with the new class(es) prepended to existing classes.
}
\description{
This is a utility function for working with the S3 class system only.
It does not work with S4 or R6 class systems.
}
\examples{
# Internal use only
obj <- add_class(list(a = 1), "my_class")
class(obj)  # c("my_class", "list")
}
\keyword{internal}
