% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_index.R
\name{analyze_single_index_model}
\alias{analyze_single_index_model}
\title{Analyze data with a single index model}
\usage{
analyze_single_index_model(
  data,
  N = 200,
  spline_seq = seq(60, 460, by = 1),
  knots = c(59, 59, 59, 59, 260, 461, 461, 461, 461),
  End = NULL,
  time_mean = NULL,
  time_sd = NULL,
  Lag_time_mean = NULL,
  Lag_time_sd = NULL
)
}
\arguments{
\item{N}{Number of independent observational units, i.e. individuals, patients, etc., for single arm of simulated data.}

\item{spline_seq}{Evaluation points over which we are conducting inference}

\item{knots}{The knots for fitting the cublic spline.}

\item{Lag_time_sd}{}
}
\description{
Analyze data with a single index model
}
