% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARC_data.R
\docType{data}
\name{ARC_data}
\alias{ARC_data}
\title{ARC Example Data}
\format{
\subsection{\code{ARC_data}}{

A data frame with 1,231 rows and 5 columns one row per visit for 301 unique subjects.
\describe{
\item{elig_pid}{Patient identifier.  301 unique values.}
\item{Asthma_control}{Outcome variable, ranges from 0 to 6.}
\item{time}{Day of visit from baseline, 0 to 1001.}
\item{Trt}{Treatment/Control group. 625 \code{control} and 606 treatment(\code{home_visits}).}
\item{visit_number}{visit number within patient, from 0 to 4}
}
}
}
\source{
TODO
}
\usage{
ARC_data
}
\description{
TODO: Description
}
\keyword{datasets}
