% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.SensIAT_fulldata_jackknife_results}
\alias{autoplot.SensIAT_fulldata_jackknife_results}
\title{Plot for estimated treatment effect for \code{SensIAT_fulldata_jackknife_results} objects}
\usage{
\method{autoplot}{SensIAT_fulldata_jackknife_results}(object, ...)
}
\description{
The horizontal and vertical axes represent the sensitivity parameter \code{alpha}
for the control and treatment groups, respectively. The plot shows
at each combination of \code{alpha} values zero if the 95\% confidence interval
contains zero, otherwise the bound of the confidence interval that is closest
to zero.
}
