% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.SensIAT_withingroup_bootstrap_results}
\alias{autoplot.SensIAT_withingroup_bootstrap_results}
\title{Plot Bootstrap Marginal Mean Curves with Confidence Bands}
\usage{
\method{autoplot}{SensIAT_withingroup_bootstrap_results}(object, time = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{SensIAT_withingroup_bootstrap_results} object.}

\item{time}{Optional vector of times. If \code{NULL}, uses an internal regular grid.}

\item{level}{Confidence level used for bootstrap intervals.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Bootstrap Marginal Mean Curves with Confidence Bands
}
