% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term2_fixed_grid.R
\name{composite_simpson}
\alias{composite_simpson}
\title{Composite Simpson's rule}
\usage{
composite_simpson(values, grid)
}
\arguments{
\item{values}{List of function values at grid points (each is a vector)}

\item{grid}{Numeric vector of grid points}
}
\value{
Integral approximation (vector of same length as each element in values)
}
\description{
Composite Simpson's rule
}
\keyword{internal}
