% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term2_fixed_grid.R
\name{composite_simpson_odd}
\alias{composite_simpson_odd}
\title{Composite Simpson's rule for odd number of points}
\usage{
composite_simpson_odd(values, grid)
}
\description{
Composite Simpson's rule for odd number of points
}
\keyword{internal}
