% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term2_fixed_grid.R
\name{composite_trapezoid}
\alias{composite_trapezoid}
\title{Composite trapezoid rule}
\usage{
composite_trapezoid(values, grid)
}
\arguments{
\item{values}{List of function values at grid points}

\item{grid}{Numeric vector of grid points}
}
\value{
Integral approximation
}
\description{
Composite trapezoid rule
}
\keyword{internal}
