% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conditional_mean_SIM.R
\name{compute_conditional_mean_SIM}
\alias{compute_conditional_mean_SIM}
\title{Compute conditional mean for single index model}
\usage{
compute_conditional_mean_SIM(alpha, X, Y, x, beta, bandwidth)
}
\arguments{
\item{bandwidth}{}
}
\value{
List with elements
\itemize{
\item \code{E(y)} and
\item \code{E(exp(alpha*y))}
}
}
\description{
Compute conditional mean for single index model
}
