% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term2_seeded_adaptive.R
\name{compute_expected_value_at_point}
\alias{compute_expected_value_at_point}
\title{Compute expected value at a single point}
\usage{
compute_expected_value_at_point(
  t,
  patient_data,
  outcome_model,
  alpha,
  impute_fn = NULL,
  time_var = NULL
)
}
\arguments{
\item{t}{Time point}

\item{patient_data}{Patient data frame}

\item{outcome_model}{Fitted outcome model}

\item{alpha}{Sensitivity parameter}

\item{impute_fn}{Imputation function}

\item{time_var}{Time variable name}
}
\value{
List with E_exp_alphaY and E_Yexp_alphaY
}
\description{
Compute expected value at a single point
}
\keyword{internal}
