% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_glm.R
\name{compute_glm_influence_term_1_for_all}
\alias{compute_glm_influence_term_1_for_all}
\title{Compute GLM Influence Term 1 for All Observations}
\usage{
compute_glm_influence_term_1_for_all(
  data,
  glm_model,
  times_all,
  ids_all,
  Y_all,
  alpha,
  intensity_weights,
  base
)
}
\description{
Compute GLM Influence Term 1 for All Observations
}
\keyword{internal}
