% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_glm.R
\name{compute_glm_influence_term_2_for_all_patients}
\alias{compute_glm_influence_term_2_for_all_patients}
\title{Compute GLM Influence Term 2 for All Patients}
\usage{
compute_glm_influence_term_2_for_all_patients(
  outcome.model,
  integration_data,
  alpha,
  base,
  id,
  time,
  tol = .Machine$double.eps^(1/3),
  ...
)
}
\description{
Compute GLM Influence Term 2 for All Patients
}
\keyword{internal}
