% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_glm.R
\name{compute_glm_influence_term_2_for_individual}
\alias{compute_glm_influence_term_2_for_individual}
\title{Compute GLM Influence Term 2 for Individual Patient}
\usage{
compute_glm_influence_term_2_for_individual(
  patient_data,
  patient_times,
  outcome.model,
  alpha,
  base,
  slope,
  tol = .Machine$double.eps^(1/3)
)
}
\description{
Compute GLM Influence Term 2 for Individual Patient
}
\keyword{internal}
