% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_influence_term_2_linearly.R
\name{compute_influence_term_2_linearly}
\alias{compute_influence_term_2_linearly}
\title{Compute Second Influence Term Assuming Piece-wise Linearity}
\usage{
compute_influence_term_2_linearly(df_i, alpha, object, base)
}
\arguments{
\item{df_i}{data frame for one individual, should include all observations including baseline.}

\item{alpha}{sensitivity parameter}

\item{object}{\code{PCORI_within_group_model} object with model data.}

\item{base}{a \link[=SplineBasis]{Spline Basis} object.}
}
\value{
A data frame with columns \code{alpha}, and \code{influence_term_2}.
The latter a list column with each element a vector.
}
\description{
Compute the second term of the influence function. defined as
$$
\int_{t=a}^b V^{-1}B(t)\hat{E}\link[=Y(t)|\\bar{O}(t)_i]{Y(t)|\bar{O}(t)_i}dt
$$
}
\details{
The limits of the integral are inferred from the knots of \code{base}.
The integral is computed on the assumption
}
