% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R, R/influence_sim.R
\name{compute_influence_terms}
\alias{compute_influence_terms}
\alias{compute_influence_terms.default}
\alias{compute_influence_terms.SensIAT::Single-index-outcome-model}
\title{Compute Influence Terms}
\usage{
compute_influence_terms(
  data_all_with_transforms,
  base,
  alpha,
  outcome.model,
  intensity.model,
  ...
)

\method{compute_influence_terms}{default}(time, base, alpha, outcome.model, intensity.model, ...)

\method{compute_influence_terms}{`SensIAT::Single-index-outcome-model`}(
  data_all_with_transforms,
  base,
  alpha,
  outcome.model,
  intensity.model,
  tolerance = .Machine$double.eps^(1/3),
  ...
)
}
\description{
Compute Influence Terms
}
\section{Methods (by class)}{
\itemize{
\item \code{compute_influence_terms(default)}: Generic method which covers lm and glm outcome models.

\item \code{compute_influence_terms(`SensIAT::Single-index-outcome-model`)}: Optimized method for the single index model.

}}
