% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_influence_function.R
\name{construct_baseline_intensity}
\alias{construct_baseline_intensity}
\title{Construct the Baseline Intensity}
\usage{
construct_baseline_intensity(visits.data, spline_seq)
}
\arguments{
\item{visits.data}{The Visits data formatted as the output from \link{formatting_fn}
\code{Visits} component.}

\item{spline_seq}{The spline sequence.}
}
\value{
The \code{visits.data} filtered to those time points that fall within the
spline_seq range with an added column \code{baseline_lambda} the subject's
baseline intensity at each of their own visit times.
}
\description{
Construct the Baseline Intensity
}
\examples{
example("formatting_fn", package='pcoriRPackage')

Visits_df <- ARC_formatted$Visits |>
    filter(, Trt == 'home_visits') |>
    construct_influence_function(60:460)
glimpse(Visits_df)
}
