% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_influence_function.R
\name{construct_influence_function}
\alias{construct_influence_function}
\title{Construct Influence Function}
\usage{
construct_influence_function(alpha)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
For a given sensitivity parameter $\alpha$, first estimate $\beta$,
then estimate $Y(t)$.
}
\details{
In the output \code{Term1_unweighted} is the estimator for
\deqn{
  \frac{
     Y_i(t_k)-\hat{E}[Y(t_k)|\overline{O}(t_k)_i]
  }{
     \hat{\rho}(t_k,Y_i(t_k),\overline{O}(t_k)_i)
  }.
}
}
