% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_intensity_model.R
\name{construct_intensity_model}
\alias{construct_intensity_model}
\title{Construct Intensity model}
\usage{
construct_intensity_model(survival.data, id.var = names(survival.data)[[1]])
}
\arguments{
\item{survival.data}{The Survival formatted data limited to the treatment group.
Must include named components \code{time}, \code{Event}, \code{Prev_time},
\code{Prev_outcome}, and \code{Visit_number}}

\item{id.var}{The patient identifier variable in \code{survival_data}}
}
\value{
A list of lists where at the top most level one list for each stratum,
and for eaach stratum a list with components \code{cumhaz}, the cummulative hazard,
and \code{base_intens}, the baseline intensity.
}
\description{
Construct Intensity model
}
\examples{
ARC_formatted <- formatting_fn(
    df              = ARC_data,
    id_var          = "elig_pid",
    treatment_var   = "Trt",
    outcome_var     = "Asthma_control",
    visitnumber_var = "Visit_number",
    time_var        = "time",
    V               = 5,
    knots           = c(59,59,59,59,260,461,461,461,461),
    spline_seq      = 60:460,
    last_day = 830
)
survival.data <- ARC_formatted$Survival |> filter(Trt == "home_visits")
intensity <- construct_intensity_model(survival_data)
glimpse(intensity)
}
