% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_curve_est.R
\name{estimate_mean_curves}
\alias{estimate_mean_curves}
\title{Title}
\usage{
estimate_mean_curves(
  data,
  patient_id_var,
  treatment,
  spline_seq,
  Knots = c(rep(min(spline_seq) - 1, 4), mean(range(spline_seq)), rep(max(spline_seq) +
    1, 4)),
  spline_fn = spline_fn,
  ...,
  End = attr(data, "End"),
  Alpha_seq = c(-0.6, -0.3, 0, 0.3, 0.6)
)
}
\arguments{
\item{Alpha_seq}{}
}
\description{
Title
}
\examples{
estimates <-
    estimate_mean_curves(
        ARC_data, elig_pid, Trt='home_visits',

    )
}
