% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term2_fixed_grid.R
\name{extract_patient_times}
\alias{extract_patient_times}
\title{Extract patient times from data}
\usage{
extract_patient_times(patient_data, time_var = NULL)
}
\arguments{
\item{patient_data}{Patient data frame}

\item{time_var}{Time variable name (optional, can be a quosure or string)}
}
\value{
Numeric vector of times
}
\description{
Extract patient times from data
}
\keyword{internal}
