% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCORI_fulldata_model.R
\name{fit_PCORI_fulldata_model}
\alias{fit_PCORI_fulldata_model}
\title{Fit the PCORI model for a single data set.}
\usage{
fit_PCORI_fulldata_model(data, trt, ...)
}
\arguments{
\item{data}{the full data set.}

\item{trt}{an expression that determine what is treated as the treatment.
Everything not treatment is considered control.}

\item{...}{Passed on to fit_within_group_model.}
}
\value{
a list with class \code{PCORI-fulldata-fitted-model} with two components,
\code{control} and \code{treatment}, each of which is an indepently fitted
\code{PCORI-within-group-fitted-model} fit with the fit_within_group_model
function.
}
\description{
Fit the PCORI model for both treatment and control groups.
}
\details{
This function is agnostic to whether it is being given the original or a
bootstrap replication.
}
