% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCORI_fulldata_model.R
\name{fit_SensIAT_fulldata_model}
\alias{fit_SensIAT_fulldata_model}
\title{Fit the sensitivity analysis model for a single data set.}
\usage{
fit_SensIAT_fulldata_model(data, trt, ...)
}
\arguments{
\item{data}{the full data set.}

\item{trt}{an expression that determine what is treated as the treatment.
Everything not treatment is considered control.}

\item{...}{Passed on to fit_within_group_model.}
}
\value{
a list with class \code{SensIAT-fulldata-fitted-model} with two components,
\code{control} and \code{treatment}, each of which is an independently fitted
\code{SensIAT-within-group-fitted-model} fit with the fit_within_group_model
function.
}
\description{
Fit the sensitivity analysis for both treatment and control groups.
}
\details{
This function is agnostic to whether it is being given the original or a
bootstrap replication.
}
\examples{
\dontrun{
model <-
    fit_SensIAT_within_group_model(
        group.data = SensIAT_example_data,
        outcome_modeler = SensIAT_sim_outcome_modeler,
        alpha = c(-0.6, -0.3, 0, 0.3, 0.6),
        id.var = Subject_ID,
        outcome.var = Outcome,
        time.var = Time,
        End = 830,
        knots = c(60,60,60,60,260,460,460,460,460),
    )
}
}
