% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term2_seeded_adaptive.R
\name{format_cache_key}
\alias{format_cache_key}
\title{Format cache key for time value}
\usage{
format_cache_key(t)
}
\arguments{
\item{t}{Numeric time value}
}
\value{
Character key
}
\description{
Format cache key for time value
}
\keyword{internal}
