% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_fn.R
\name{formatting_fn}
\alias{formatting_fn}
\title{Format the original ARC data}
\usage{
formatting_fn(
  df,
  id_var,
  treatment_var,
  visitnumber_var,
  outcome_var,
  time_var,
  spline_seq,
  V,
  knots,
  p,
  last_day = max(df[[time_var]], na.rm = TRUE)
)
}
\arguments{
\item{df}{cleaned ARC data}

\item{id_var}{A string indicating the patient/subject identifier variable, must be present in \code{df}.}

\item{treatment_var}{A string indicating the variable containing the treatment group, , must be present in \code{df}.}

\item{visitnumber_var}{A string indicating an numeric(integer) variable from \code{df} to identify the visit number within patient.}

\item{outcome_var}{A string indicating the outcome variable from \code{df}.}

\item{time_var}{A string indicating the time variable from \code{df}.}

\item{spline_seq}{passed to \link{spline_fn} to create the spline coefficient matrix.}

\item{V}{TODO\link{@yujinggao24}: please clarify.}

\item{knots}{passed on in output.}

\item{p}{The number of rows of data.  TODO\link{@yujinggao24}: Please Verify.}

\item{last_day}{Last time point of observation, defaults to the maximum time present in \code{df}.}
}
\value{
A list of three \href{data.frame}{data frames} with class \code{arc}:
\itemize{
\item baseline visits,
\item post-baseline visits, and
\item a \href{data.frame}{data frame} for use with the \href{package:survival}{\code{survival}}
package, which has rows indicationg censoring for subjects with fewer than
4 post-baseline assessments.
}

Also attached are attributes:
\itemize{
\item spline_seq, the same as input;
\item End, equal to input \code{last_day};
\item V_inverse, $V^{-1}$ where $V=B(t)B(t)'$ and $B(t)$ is the spline matrix created by
applying \code{spline_fn} to \code{spline_seq};
\item Weights_term2, $V^{-1}B(t)$;
\item knots, passed through from input.
\item p, the number of rows of $B(t)$.
}
}
\description{
This performs additional formatting on the cleaned ARC data.
}
