% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_post_baseline_times.R
\name{generate_post_baseline_times}
\alias{generate_post_baseline_times}
\title{Function to generate times of a given post-baseline assessment}
\usage{
generate_post_baseline_times(
  Start_vec,
  Visit = 1,
  Outcomes_vec,
  lambda0_t,
  lambda_star,
  cox_model
)
}
\arguments{
\item{cox_model}{}
}
\description{
Uses Ogata's Thinning Algorithm to generate times that will have
a prescribed intensity.
}
