% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda0_fn.R
\name{lambda0_fn}
\alias{lambda0_fn}
\title{Kernel-smooth a cumulative intensity using an Epanechnikov kernel}
\usage{
lambda0_fn(t, b, surv)
}
\arguments{
\item{t}{time, a single time point.}

\item{b}{bandwidth.}

\item{surv}{a \code{\link[survival:survfit]{survival::survfit}} object, generated by the survival package.}
}
\value{
The kernel-smoothed cumulative intensity at time \code{t}.
}
\description{
Kernel-smooth a cumulative intensity using an Epanechnikov kernel
}
\examples{
fit <- survival::survfit(survival::Surv(time, status) ~ x, data = survival::aml)
sapply(15:45, lambda0_fn, b=10, fit)
}
