% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{pcori_jackknife}
\alias{pcori_jackknife}
\title{Estimate response with jackknife resampling}
\usage{
pcori_jackknife(original.object, time, ...)
}
\arguments{
\item{original.object}{A PCORI_within_group_model object.}

\item{time}{Time points for which to estimate the response.}

\item{...}{currently ignored.}
}
\value{
A tibble with columns alpha, time, jackknife_mean, and jackknife_var,
where jackknife_mean is the mean of the jackknife estimates and jackknife_var
is the estimated variances of the response at the given timepoints for the
specified alpha values.
}
\description{
Estimate response with jackknife resampling
}
\examples{
original.object <-
fit_PCORI_within_group_model(
    group.data = PCORI_example_data,
    outcome_modeler = PCORI_sim_outcome_modeler,
    alpha = c(-0.6, -0.3, 0, 0.3, 0.6),
    id.var = Subject_ID,
    outcome.var = Outcome,
    time.var = Time,
    intensity.bandwidth = 30,
    knots = c(60,60,60,60,260,460,460,460,460),
    End = 830
)
jackknife.estimates <- pcori_jackknife(original.object, time = c(90, 180, 270, 360, 450))
}
