% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbm_outcome_modeler.R
\name{PCORI_nb_outcome_modeler}
\alias{PCORI_nb_outcome_modeler}
\title{Outcome Modeler for PCORI Negative Binomial Model.}
\usage{
PCORI_nb_outcome_modeler(formula, data, ...)
}
\arguments{
\item{formula}{The outcome model formula}

\item{data}{The data to fit the outcome model to.}

\item{...}{Currently ignored, included for future compatibility.}
}
\value{
Object of class \code{PCORI::Negative-binomial-model} which contains the outcome model portion.
}
\description{
Outcome Modeler for PCORI Negative Binomial Model.
}
