% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_modeler.R
\name{PCORI_sim_outcome_modeler}
\alias{PCORI_sim_outcome_modeler}
\title{Outcome Modeler for PCORI Single Index Model.}
\usage{
PCORI_sim_outcome_modeler(formula, data, kernel = "dnorm", method = "nmk", ...)
}
\arguments{
\item{formula}{The outcome model formula}

\item{data}{The data to fit the outcome model to.}

\item{...}{Currently ignored, included for future compatibility.}
}
\value{
TODO
}
\description{
Outcome Modeler for PCORI Single Index Model.
}
