% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_compute_influence_term_2_quadv_sim_via_matrix}
\alias{pcoriaccel_compute_influence_term_2_quadv_sim_via_matrix}
\title{Runs an optimized implementation of the \code{compute_influence_term_2_quadv_sim_via_matrix}
function.}
\usage{
pcoriaccel_compute_influence_term_2_quadv_sim_via_matrix(
  X,
  Y,
  times,
  individual_X,
  x_slope,
  alpha,
  beta,
  spline_basis,
  bandwidth,
  tol = 0.0001220703,
  kernel = "K2_Biweight"
)
}
\arguments{
\item{X}{Matrix of all covariates, transformed as necessary by model}

\item{Y}{Vector of all outcomes (same length as a column of \code{X})}

\item{times}{Vector of observation times for individual}

\item{individual_X}{Matrix of covariates for individual rows correspond to times prepared for
inferences for integration.}

\item{x_slope}{Vector of numeric(length(beta)) indicating how}

\item{alpha}{Vector of sensitivity parameters}

\item{beta}{Vector of coefficients of the outcome model}

\item{spline_basis}{Spline basis object (\code{orthogonalsplinebasis::SplineBasis})}

\item{bandwidth}{Bandwidth for the kernel density estimate of the outcome model.}

\item{tol}{Tolerance for integration}

\item{kernel}{Kernel function to use for the kernel density estimate}
}
\value{
integration result
}
\description{
Runs an optimized implementation of the \code{compute_influence_term_2_quadv_sim_via_matrix}
function.
}
\keyword{internal}
