% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_estimate_pmf}
\alias{pcoriaccel_estimate_pmf}
\title{Directly estimate the probability mass function of Y.}
\usage{
pcoriaccel_estimate_pmf(Xb, Y, xi, y_seq, h, kernel = "K2_Biweight")
}
\arguments{
\item{Xb}{Numeric vector of individual linear predictors from the data}

\item{Y}{Numeric vector of individual responses from the data}

\item{xi}{value of the individuals linear predictor at the point of estimation}

\item{y_seq}{Numeric vector of unique values of \code{Y}.}

\item{h}{bandwidth of the kernel}

\item{kernel}{character string specifying the kernel to use, either \code{"dnorm"}, \code{"K2_Biweight"}, or \code{"K4_Biweight"}}
}
\description{
Directly estimate the probability mass function of Y.
}
