% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_evaluate_basis}
\alias{pcoriaccel_evaluate_basis}
\title{Rcpp version of \verb{evaluate_basis(⋯)} function}
\usage{
pcoriaccel_evaluate_basis(spline_basis, x)
}
\arguments{
\item{spline_basis}{The spline basis, S4 class \code{orthogonalsplinebasis::SplineBasis}}

\item{x}{The point to evaluate}
}
\value{
Vector fyxb
}
\description{
Rcpp version of \verb{evaluate_basis(⋯)} function
}
