% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_evaluate_basis_mat}
\alias{pcoriaccel_evaluate_basis_mat}
\title{Compiled version of \code{evaluate_basis()} function (matrix version)}
\usage{
pcoriaccel_evaluate_basis_mat(spline_basis, x)
}
\arguments{
\item{spline_basis}{The spline basis, S4 class \code{orthogonalsplinebasis::SplineBasis}}

\item{x}{numeric vector of points to evaluate}
}
\value{
Matrix of the basis functions evaluated at x.
}
\description{
Compiled version of \code{evaluate_basis()} function (matrix version)
}
