% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_hello}
\alias{pcoriaccel_hello}
\title{Returns a string, just as a basic check that the C++ plugin library is working.}
\usage{
pcoriaccel_hello()
}
\value{
hello string
}
\description{
Returns a string, just as a basic check that the C++ plugin library is working.
}
