% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_inner_prod}
\alias{pcoriaccel_inner_prod}
\title{Inner product (dot product) of two vectors.}
\usage{
pcoriaccel_inner_prod(vecA, vecB)
}
\arguments{
\item{vecA}{first vector}

\item{vecB}{second vector}
}
\value{
vecAᵀ * vecB = vecA • vecB
}
\description{
Inner product (dot product) of two vectors.
}
