% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_integrate_simp}
\alias{pcoriaccel_integrate_simp}
\title{Integrate function using adaptive Simpson quadrature.}
\usage{
pcoriaccel_integrate_simp(integrand, lo, hi, tol = 1.490116e-08)
}
\arguments{
\item{integrand}{The integrand, must take scalar argument, may return scalar, vector, or matrix.}

\item{lo}{Lower integration bound}

\item{hi}{Upper integration bound}

\item{tol}{Tolerance for integration, default \code{.Machine$double.eps^(1/2)}}
}
\value{
integration result, list with elements \verb{$Q} (the integral estimate), \verb{$fcnt} (the number
of function evaluations), and \verb{$estim.prec} (a (pessimistic) estimate of the precision).
}
\description{
Integrate function using adaptive Simpson quadrature.
}
\keyword{internal}
