% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_mmul}
\alias{pcoriaccel_mmul}
\title{Multiplies two matrices.  If the first argument is a vector, it is interpreted as a row vector.
Otherwise, if the second argument is a vector, it is interpreted as a column vector.}
\usage{
pcoriaccel_mmul(matrA, matrB)
}
\arguments{
\item{matrA}{first matrix}

\item{matrB}{second matrix}
}
\value{
\code{matrA * matrB}
}
\description{
Multiplies two matrices.  If the first argument is a vector, it is interpreted as a row vector.
Otherwise, if the second argument is a vector, it is interpreted as a column vector.
}
