% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_NW_basic}
\alias{pcoriaccel_NW_basic}
\title{Runs a \emph{basic} implementation of the "NW" function with the "K2_Biweight" kernel, just as a
proof-of-concept.}
\usage{
pcoriaccel_NW_basic(Xb, Y, xb, y_seq, h)
}
\arguments{
\item{Xb}{a vector (expected to be about 500 elements)}

\item{Y}{a vector (same size as \code{Xb})}

\item{xb}{a vector}

\item{y_seq}{a vector}

\item{h}{a scalar, the bandwidth of kernel}
}
\value{
A matrix of the same size as \code{xb} by \code{y_seq}.
}
\description{
Runs a \emph{basic} implementation of the "NW" function with the "K2_Biweight" kernel, just as a
proof-of-concept.
}
\keyword{internal}
