% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_outer_prod}
\alias{pcoriaccel_outer_prod}
\title{Outer product of two vectors.}
\usage{
pcoriaccel_outer_prod(vecA, vecB)
}
\arguments{
\item{vecA}{first vector}

\item{vecB}{second vector}
}
\value{
vecA * vecBᵀ = vecA ⊗ vecB
}
\description{
Outer product of two vectors.
}
\examples{
pcoriaccel_outer_prod( c(1,2,3,4,5), c(2,4,6) )
}
