% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_sorted_unique}
\alias{pcoriaccel_sorted_unique}
\title{Returns the unique elements of a vector, sorted in ascending order.}
\usage{
pcoriaccel_sorted_unique(vec)
}
\arguments{
\item{vec}{the vector}
}
\value{
sort(unique(vec))
}
\description{
Returns the unique elements of a vector, sorted in ascending order.
}
