% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCORI_within_group_model.R
\name{predict.PCORI_within_group_model}
\alias{predict.PCORI_within_group_model}
\title{Prediction method for \code{PCORI_within_group_model} objects}
\usage{
\method{predict}{PCORI_within_group_model}(object, time, alpha, spline_fn, spline_seq, knots)
}
\arguments{
\item{object}{a \code{PCORI_within_group_model} object.}

\item{time}{The time(s) to evaluate at.}

\item{alpha}{The alpha values to evaluate at.

Evaluate the fitted model, \code{object}, at each combination of \code{time} and
\code{alpha}.}
}
\value{
A tibble/data.frame with the components: \code{time}, \code{alpha}, \code{mean}, \code{var}.
where \code{time} and \code{alpha} are the combinations of the respective input(s) and
\code{mean} and \code{var} are the estimated mean and variance of the response for the
given model.
}
\description{
Prediction method for \code{PCORI_within_group_model} objects
}
\examples{

fitted.trt.nb <-
    fit_PCORI_within_group_model(
        group.data = filter(ARC_data, Trt=='home_visits'),
        id.var = elig_pid,
        outcome.var = Asthma_control,
        time.var = time,
        intensity.bandwidth = 30,
        End = 830,
        outcome_modeler = glm.nb
    )
predict(fitted.trt.nb, time = c(90, 180), alpha = c(-0.6, -0.3, 0, 0.3, 0.6)
       , spline_fn, spline_seq=seq(60, 460, by=1), knots=c(59,59,59,59,260,461,461,461,461)
)

fitted.trt.sim <-
    fit_PCORI_within_group_model(
        group.data = filter(ARC_data, Trt=='home_visits'),
        outcome_modeler = PCORI_sim_outcome_modeler,
        id.var = elig_pid,
        outcome.var = Asthma_control,
        time.var = time,
        intensity.bandwidth = 30,
        End = 830
    )
predict(fitted.trt.sim, time = c(90, 180),
        alpha = c(-0.6, -0.3, 0, 0.3, 0.6),
        spline_fn, spline_seq=seq(60, 460, by=1), knots=c(59,59,59,59,260,461,461,461,461)
)
}
