% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCORI_fulldata_model.R
\name{predict.SensIAT_fulldata_model}
\alias{predict.SensIAT_fulldata_model}
\title{Prediction method for \code{SensIAT_fulldata_model} objects.}
\usage{
\method{predict}{SensIAT_fulldata_model}(object, time, ...)
}
\arguments{
\item{object}{an object of class \code{SensIAT_fulldata_model}}

\item{time}{time points to evaluate at}

\item{...}{additional arguments passed to \code{predict.SensIAT_within_group_model}}
}
\value{
a \code{tibble}/\code{data.frame} with the following components.
\itemize{
\item \code{time}, the time point
}
}
\description{
For each combination of \code{time} and \code{alpha} estimate the mean response and
variance for each group as well as estimate the mean treatment effect and
variance.
}
\examples{
\dontrun{
model <-
    fit_SensIAT_within_group_model(
        group.data = SensIAT_example_data,
        outcome_modeler = SensIAT_sim_outcome_modeler,
        alpha = c(-0.6, -0.3, 0, 0.3, 0.6),
        id.var = Subject_ID,
        outcome.var = Outcome,
        time.var = Time,
        End = 830,
        knots = c(60,60,60,60,260,460,460,460,460),
    )
predict(model, time = c(90, 180))
}
}
