% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_within_group.R
\name{predict.SensIAT_withingroup_bootstrap_results}
\alias{predict.SensIAT_withingroup_bootstrap_results}
\title{Predict Marginal Mean from Bootstrap Coefficient Replicates}
\usage{
\method{predict}{SensIAT_withingroup_bootstrap_results}(object, time, include.var = TRUE, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{SensIAT_withingroup_bootstrap_results} object.}

\item{time}{Time points of interest.}

\item{include.var}{Logical. If \code{TRUE} and link is identity, include the
original-model asymptotic variance column \code{var}.}

\item{level}{Confidence level used to produce \code{lower} and \code{upper} bounds.}

\item{...}{Currently ignored.}
}
\value{
A \code{tibble} with bootstrap summaries by \code{alpha} and \code{time}.
}
\description{
Predict Marginal Mean from Bootstrap Coefficient Replicates
}
