% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SensIAT::Single-index-outcome-model}
\alias{print.SensIAT::Single-index-outcome-model}
\title{Print method for \code{SensIAT::Single-index-outcome-model}}
\usage{
\method{print}{`SensIAT::Single-index-outcome-model`}(x, ...)
}
\arguments{
\item{x}{A \code{SensIAT::Single-index-outcome-model} object.}

\item{...}{Currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{SensIAT::Single-index-outcome-model}
}
