% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SensIAT_marginal_mean_model_generalized}
\alias{print.SensIAT_marginal_mean_model_generalized}
\title{Print method for \code{SensIAT_marginal_mean_model_generalized}}
\usage{
\method{print}{SensIAT_marginal_mean_model_generalized}(x, ...)
}
\arguments{
\item{x}{A \code{SensIAT_marginal_mean_model_generalized} object.}

\item{...}{Currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{SensIAT_marginal_mean_model_generalized}
}
