% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SensIAT_within_group_model}
\alias{print.SensIAT_within_group_model}
\title{Print method for \code{SensIAT_within_group_model}}
\usage{
\method{print}{SensIAT_within_group_model}(x, ...)
}
\arguments{
\item{x}{A \code{SensIAT_within_group_model} object.}

\item{...}{Currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{SensIAT_within_group_model}
}
