% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_bootstrap.R
\name{sample_parametric_coeffs}
\alias{sample_parametric_coeffs}
\title{Sample parametric coefficients from a fitted model using asymptotic MVN}
\usage{
sample_parametric_coeffs(model)
}
\arguments{
\item{model}{A fitted model with \code{coef()} and \code{vcov()} methods}
}
\value{
Named numeric vector of sampled coefficients
}
\description{
Sample parametric coefficients from a fitted model using asymptotic MVN
}
