% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensIAT-package.R
\docType{package}
\name{SensIAT-package}
\alias{SensIAT}
\alias{SensIAT-package}
\title{SensIAT: Sensitivity Analysis for Irregular Assessment Times}
\description{
Sensitivity analysis for trials with irregular and informative assessment times, based on a new influence function-based, augmented inverse intensity-weighted estimator.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofUEpiBio/SensIAT}
  \item Report bugs at \url{https://github.com/UofUEpiBio/SensIAT/issues}
}

}
\author{
\strong{Maintainer}: Andrew Redd \email{andrew.redd@hsc.utah.edu} (\href{https://orcid.org/0000-0002-6149-2438}{ORCID})

Authors:
\itemize{
  \item Yujing Gao \email{ygao39@ncsu.edu}
  \item Shu Yang \email{syang24@ncsu.edu}
  \item Bonnie Smith \email{bsmit179@jhmi.edu}
  \item Daniel Scharfstein \email{daniel.scharfstein@hsc.utah.edu} (\href{https://orcid.org/0000-0001-7482-9653}{ORCID}) [project director]
}

Other contributors:
\itemize{
  \item Agatha Mallett \email{agatha@geometrian.com} [contributor, contractor]
  \item University of Utah [copyright holder]
}

}
\keyword{internal}
