% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_marginal_model.R
\name{SensIAT_fit_marginal_model}
\alias{SensIAT_fit_marginal_model}
\title{Title}
\usage{
SensIAT_fit_marginal_model(
  data,
  alpha,
  knots,
  intensity.model,
  outcome.model,
  spline.degree = 3L,
  ...
)
}
\arguments{
\item{data}{Data for evaluation of the model. Should match the data used to fit the intensity and outcome models.}

\item{alpha}{Sensitivity parameter, a vector of values.}

\item{knots}{Location of spline knots. If a \code{SplineBasis} object is provided, it is used directly.}

\item{intensity.model}{The assessment time intensity model.}

\item{outcome.model}{The observed effects model.}

\item{spline.degree}{The degree of the spline basis, default is 3 (cubic splines).}

\item{...}{Additional arguments passed to \code{compute_influence_terms}.}
}
\value{
a list with the fitted model, including the coefficients and their variances for each alpha value.
}
\description{
Title
}
