% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_expected_values.R
\name{sensitivity_expected_values}
\alias{sensitivity_expected_values}
\title{Compute Conditional Expected Values based on Outcome Model}
\usage{
sensitivity_expected_values(
  model,
  alpha = 0,
  new.data = model.frame(model),
  ...
)
}
\arguments{
\item{model}{An object representing the output of the outcome model.}

\item{alpha}{The sensitivity parameter}

\item{new.data}{Data to compute conditional means for, defaults to the model frame for the fitted model.}

\item{...}{passed onto methods.}
}
\value{
The \code{new.data} frame with additional columns \code{E_Yexp_alphaY}, and \code{E_exp_alphaY} appended.
}
\description{
Compute Conditional Expected Values based on Outcome Model
}
\details{
Compute the conditional expectations needed for predictions in the models.
Two additional values/expectations are computed:
\itemize{
\item \verb{$E \\big[ Y(t) \\exp \\\{  \\alpha Y(t) \\\}   | A(t)=1, \\bar\{O\}(t) \\big]$}, returned as \code{E_Yexp_alphaY}, and
\item \verb{$E \\big[ \\exp \\\{ \\alpha Y(t) \\\} \\  | A(t)=1, \\bar\{O\}(t) \\big]$}, returned as \code{E_exp_alphaY}.
}
}
