% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_modeler.R
\name{SIDRnew}
\alias{SIDRnew}
\title{Multi-index distribution regression}
\usage{
SIDRnew(
  X,
  Y,
  Y.CP = NULL,
  initial = NULL,
  kernel = "K2_Biweight",
  method = "optim",
  optim_method = "BFGS",
  abs.tol = 1e-04,
  bandwidth = NULL,
  wi.boot = NULL
)
}
\arguments{
\item{kernel}{It could be any of
\code{K2_Biweight}, \code{dnorm}, \code{K4_Biweight}, \code{K2_Biweight}, or \code{K4_Biweight}}

\item{optim_method}{It could be any of
\code{Nelder-Mead}, \code{BFGS}, \code{CG}, \code{L-BFGS-B}, or \code{Brent}}

\item{abs.tol}{}
}
\description{
Multi-index distribution regression
}
