% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRnew.R
\name{SIDRnew_fixed_bandwidth}
\alias{SIDRnew_fixed_bandwidth}
\title{Fit Single Index Dimension Reduction with Fixed Bandwidth}
\usage{
SIDRnew_fixed_bandwidth(
  X,
  Y,
  ids,
  Y.CP = NULL,
  initial = NULL,
  kernel = "K2_Biweight",
  method = c("optim", "nlminb", "nmk"),
  optim_method = "BFGS",
  abs.tol = 1e-04
)
}
\description{
Fit Single Index Dimension Reduction with Fixed Bandwidth
}
\keyword{internal}
