% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_class}
\alias{add_class}
\title{Adds an S3 class to an object}
\usage{
add_class(x, class)
}
\arguments{
\item{x}{An object to which the class should be added.}

\item{class}{A character vector of class names to be added.}
}
\description{
Adds an S3 class to an object
}
\examples{
add_class(TRUE, 'flag')
}
