% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.SensIAT_fulldata_jackknife_results}
\alias{autoplot.SensIAT_fulldata_jackknife_results}
\title{Plot for estimated treatment effect for \code{SensIAT_fulldata_jackknife_results} objects}
\usage{
\method{autoplot}{SensIAT_fulldata_jackknife_results}(object, ..., include.rugs = NA)
}
\arguments{
\item{object}{A \code{SensIAT_fulldata_jackknife_results} object.}

\item{...}{Additional arguments passed to \code{predict}.}

\item{include.rugs}{If \code{TRUE}, adds rugs to the plot. If \code{FALSE}, no rugs are added.
When \code{NA}, rugs are added only if the number of distinct values of \code{alpha_control}
and \code{alpha_treatment} is less than or equal to 10.}
}
\description{
The horizontal and vertical axes represent the sensitivity parameter \code{alpha}
for the control and treatment groups, respectively. The plot shows
at each combination of \code{alpha} values zero if the 95\% confidence interval
contains zero, otherwise the bound of the confidence interval that is closest
to zero.
}
