% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conditional_mean_SIM.R
\name{compute_conditional_mean_SIM}
\alias{compute_conditional_mean_SIM}
\title{Compute conditional mean for single index model}
\usage{
compute_conditional_mean_SIM(alpha, X, Y, x, beta, bandwidth)
}
\arguments{
\item{alpha}{The sensitivity parameter value.}

\item{X}{The design matrix.}

\item{Y}{The outcome.}

\item{x}{The covariate vector for the observation of interest.}

\item{beta}{The coefficient vector.}

\item{bandwidth}{The bandwidth for the kernel.}
}
\value{
List with elements
\itemize{
\item \eqn{E(y)} and
\item \eqn{E(exp(alpha*y))}
}
}
\description{
Compute conditional mean for single index model
}
