% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_influence_term_2_linearly.R
\name{compute_influence_term_2_linearly}
\alias{compute_influence_term_2_linearly}
\title{Compute Second Influence Term Assuming Piece-wise Linearity}
\usage{
compute_influence_term_2_linearly(
  df_i,
  expected_value,
  base,
  variables,
  alpha,
  ...
)
}
\arguments{
\item{df_i}{data frame for one individual, should include all observations including baseline.}

\item{expected_value}{The function to compute the expected value of the outcome model.}

\item{base}{a \link[=SplineBasis]{Spline Basis} object.}

\item{variables}{a variables mapping.}

\item{alpha}{the sensitivity parameter.}

\item{...}{passed to \link{impute_patient_df}.}
}
\value{
A data frame with columns \code{alpha}, and \code{influence_term_2}.
The latter a list column with each element a vector.
}
\description{
Compute the second term of the influence function. defined as
\deqn{
\int_{t=a}^b V^{-1}B(t)\hat{E}[Y(t)|\bar{O}(t)_i]dt
}
}
\details{
The limits of the integral are inferred from the knots of \code{base}.
The integral is computed on the assumption
}
\keyword{internal}
