% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SensIAT_marginal_mean_model_generalized.R
\name{fit_SensIAT_marginal_mean_model_generalized}
\alias{fit_SensIAT_marginal_mean_model_generalized}
\title{Fit the marginal mean model for generalize outcomes.}
\usage{
fit_SensIAT_marginal_mean_model_generalized(
  data,
  id,
  alpha,
  knots,
  outcome.model,
  intensity.model,
  loss = c("lp_mse", "mean_mse", "quasi-likelihood"),
  link = c("identity", "log", "logit"),
  spline.degree = 3L,
  ...
)
}
\arguments{
\item{data}{Data for evaluation of the model. Should match the data used to fit the intensity and outcome models.}

\item{id}{The subject identifier variable in the data. Lazy evaluation is used, so it can be a symbol or a string.}

\item{alpha}{Sensitivity parameter, a vector of values.}

\item{knots}{Location of spline knots. If a \code{SplineBasis} object is provided, it is used directly.}

\item{outcome.model}{The observed effects model.}

\item{intensity.model}{The assessment time intensity model.}

\item{loss}{The loss function to use. Options are "lp_mse", "mean_mse", and "quasi-likelihood".}

\item{link}{The link function to use. Options are "identity", "log", and "logit".}

\item{spline.degree}{The degree of the spline basis, default is 3 (cubic splines).}

\item{...}{Additional arguments passed to \code{compute_influence_terms}.}
}
\description{
Fit the marginal mean model for generalize outcomes.
}
\examples{
data_with_lags <- SensIAT_example_data |>
    dplyr::group_by(Subject_ID) |>
    dplyr::mutate(
        ..prev_outcome.. = dplyr::lag(Outcome, default = NA_real_, order_by = Time),
        ..prev_time.. = dplyr::lag(Time, default = 0, order_by = Time),
        ..delta_time.. = Time - dplyr::lag(.data$Time, default = NA_real_, order_by = Time)
    )

# Create the observation time intensity model
intensity.model <-
    coxph(Surv(..prev_time.., Time, !is.na(Outcome)) ~ ..prev_outcome.. + strata(Visit),
    data = data_with_lags |> dplyr::filter(.data$Time > 0))

# Create the observed outcome model
outcome.model <-
    fit_SensIAT_single_index_fixed_coef_model(
        Outcome ~ ns(..prev_outcome.., df=3) + ..delta_time.. - 1,
        id = Subject_ID,
        data = data_with_lags |> filter(Time > 0))
fit_SensIAT_marginal_mean_model_generalized(
   data = data_with_lags, Subject_ID,
   alpha = 0,
   knots = c(60, 260, 460),
   outcome.model = outcome.model,
   intensity.model = intensity.model,
   loss = 'lp_mse',
   link = 'log')
impute_data = \(t, df){
    data_wl <- df |>
        mutate(..prev_time.. = Time,
               ..prev_outcome.. = Outcome,
               ..delta_time.. = 0)
    extrapolate_from_last_observation(t, data_wl, 'Time', slopes = c('..delta_time..' = 1))
}
time = data_with_lags$Time
id = data_with_lags$Subject_ID

}
