% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_bootstrap.R
\name{get_bootstrap_coeffs}
\alias{get_bootstrap_coeffs}
\title{Get coefficients for bootstrap simulation.}
\usage{
get_bootstrap_coeffs(model, sample_coefficients = FALSE, model_label = "model")
}
\description{
By default this returns original fitted coefficients. If sampling is requested,
coefficients are sampled from an asymptotic MVN only when \code{vcov()} is available.
}
