% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\alias{jackknife.SensIAT_within_group_model}
\alias{jackknife.SensIAT_fulldata_model}
\title{Perform Jackknife resampling on an object.}
\usage{
jackknife(object, ...)

\method{jackknife}{SensIAT_within_group_model}(object, ...)

\method{jackknife}{SensIAT_fulldata_model}(object, ...)
}
\arguments{
\item{object}{An object to cross validate on.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A data frame of the jackknife resampling results.
}
\description{
Perform Jackknife resampling on an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{jackknife(SensIAT_within_group_model)}: Perform jackknife resampling on a \code{SensIAT_within_group_model} object.

\item \code{jackknife(SensIAT_fulldata_model)}: Perform jackknife resampling on a \code{SensIAT_fulldata_model} object.

}}
