% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda0_fn.R
\name{lambda0_fn}
\alias{lambda0_fn}
\title{Kernel-smooth a cumulative intensity using an Epanechnikov kernel}
\usage{
lambda0_fn(t, b, surv)
}
\arguments{
\item{t}{time, a single time point.}

\item{b}{bandwidth.}

\item{surv}{a \href{survival:survfit}{survfit} object, generated by the survival package.}
}
\description{
Kernel-smooth a cumulative intensity using an Epanechnikov kernel
}
\examples{
fit <- survival::survfit(Surv(time, status) ~ x, data = survival::aml)
sapply(15:45, lambda0_fn, b=10, fit)
}
