% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_modeler.R
\name{PCORI_sim_outcome_modeler}
\alias{PCORI_sim_outcome_modeler}
\title{Outcome Modeler for PCORI Single Index Model.}
\usage{
PCORI_sim_outcome_modeler(
  formula,
  data,
  kernel = "K2_Biweight",
  method = "nmk",
  ...
)
}
\arguments{
\item{formula}{The outcome model formula}

\item{data}{The data to fit the outcome model to.}

\item{kernel}{The kernel to use for the outcome model.}

\item{method}{The optimization method to use for the outcome model, either "optim", "nlminb", or "nmk".}

\item{...}{Currently ignored, included for future compatibility.}
}
\value{
Object of class \code{PCORI::Single-index-outcome-model} which contains the outcome model portion.
}
\description{
Outcome Modeler for PCORI Single Index Model.
}
