% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_estimate_pmf}
\alias{pcoriaccel_estimate_pmf}
\title{Estimate the PMF directly with the K2_Biweight kernel.}
\usage{
pcoriaccel_estimate_pmf(X, Y, xi, y_seq, h)
}
\arguments{
\item{Y}{vector (same size as Xb)}

\item{y_seq}{vector}

\item{h}{scalar bandwidth of kernel}

\item{Xb}{vector (expected to be about 500 elements)}

\item{xb}{vector}
}
\value{
estimated PMF
}
\description{
Estimate the PMF directly with the K2_Biweight kernel.
}
