% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_estimate_pmf}
\alias{pcoriaccel_estimate_pmf}
\title{Estimate the PMF directly with the K2_Biweight kernel.}
\usage{
pcoriaccel_estimate_pmf(X, Y, xi, y_seq, h)
}
\arguments{
\item{Y}{a vector (same size as Xb)}

\item{xi}{a scalar}

\item{y_seq}{a vector}

\item{h}{a scalar, the bandwidth of kernel}

\item{Xb}{a vector (expected to be about 500 elements)}
}
\value{
estimated PMF
}
\description{
Estimate the PMF directly with the K2_Biweight kernel.
}
