% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_evaluate_basis}
\alias{pcoriaccel_evaluate_basis}
\title{Compiled version of \code{evaluate_basis()} function}
\usage{
pcoriaccel_evaluate_basis(spline_basis, x)
}
\arguments{
\item{spline_basis}{The spline basis, S4 class \code{orthogonalsplinebasis::SplineBasis}}

\item{x}{The point to evaluate}
}
\value{
Vector of the basis functions evaluated at x.
}
\description{
Compiled version of \code{evaluate_basis()} function
}
