% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_NW}
\alias{pcoriaccel_NW}
\title{Runs an optimized implementation of the "NW" function.}
\usage{
pcoriaccel_NW(Xb, Y, xb, y_seq, h, kernel = "K2_Biweight")
}
\arguments{
\item{Xb}{vector (expected to be about 500 elements)}

\item{Y}{vector (same size as Xb)}

\item{xb}{vector}

\item{y_seq}{vector}

\item{h}{scalar bandwidth of kernel}
}
\value{
Matrix fyxb
}
\description{
Runs an optimized implementation of the "NW" function.
}
