% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_NW}
\alias{pcoriaccel_NW}
\title{Runs an optimized implementation of the "NW" function.}
\usage{
pcoriaccel_NW(Xb, Y, xb, y_seq, h, kernel = "K2_Biweight")
}
\arguments{
\item{Xb}{a vector (expected to be about 500 elements)}

\item{Y}{a vector (same size as Xb)}

\item{xb}{a vector}

\item{y_seq}{a vector}

\item{h}{a scalar, the bandwidth of kernel}
}
\value{
Matrix fyxb
}
\description{
Runs an optimized implementation of the "NW" function.
}
