% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcoriaccel_outer_sum}
\alias{pcoriaccel_outer_sum}
\title{Outer sum of two vectors.}
\usage{
pcoriaccel_outer_sum(vecA, vecB)
}
\arguments{
\item{vecA}{first vector}

\item{vecB}{second vector}
}
\value{
\verb{vecA ⊕ vecB}
}
\description{
Outer sum of two vectors.
}
\keyword{internal}
