% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCORI_within_group_model.R
\name{predict.PCORI_within_group_model}
\alias{predict.PCORI_within_group_model}
\title{Prediction method for \code{PCORI_within_group_model} objects}
\usage{
\method{predict}{PCORI_within_group_model}(
  object,
  time,
  alpha,
  knots,
  intensity.bandwidth = NULL,
  integral.resolution = 1000,
  integration.method = c("linear", "numerical", "piecewise", "quadv"),
  ...
)
}
\arguments{
\item{object}{a \code{PCORI_within_group_model} object.}

\item{time}{The time(s) to evaluate at.}

\item{alpha}{The values for the sensitivity parameter to evaluate at.}

\item{knots}{The knots used for the spline function.
Includes both boundary and internal knots.}

\item{integral.resolution}{the number of points to use for numerical integration.

Evaluate the fitted model, \code{object}, at each combination of \code{time} and
\code{alpha}.}

\item{integration.method}{Method for integration when computing the second influence term.  See Details.}

\item{spline_fn}{Function to to generate the values for the spline functions
with knots given by \code{knots}.  Must accept time as the first argument and
knots as the second.}

\item{spline_seq}{The time points used to approximate the integral in term 2
of the influence function.}
}
\value{
A tibble/data.frame with the components: \code{time}, \code{alpha}, \code{mean}, \code{var}.
where \code{time} and \code{alpha} are the combinations of the respective input(s) and
\code{mean} and \code{var} are the estimated mean and variance of the response for the
given model.

@details
For \code{integration.method} when computing the integral in term two of the
influence function \code{linear} approximates the expected value as piece-wise
linear and is much faster, \code{numerical} uses traditional numerical
integration.
}
\description{
Prediction method for \code{PCORI_within_group_model} objects
}
\examples{

fitted.trt.sim <-
    fit_PCORI_within_group_model(
        group.data = PCORI_example_data,
        outcome_modeler = PCORI_sim_outcome_modeler,
        id.var = Subject_ID,
        outcome.var = Outcome,
        time.var = Time,
        End = 830
    )
time.pw <- system.time({
pred.pw <- predict(fitted.trt.sim, time = c(90, 180),
    alpha = c(-0.6, -0.3, 0, 0.3, 0.6),
    intensity.bandwidth = 30,
    knots=c(60,60,60,60,260,460,460,460,460),
    integration.method = 'piecewise'
)
})
time.num <- system.time({
pred.num <- predict(fitted.trt.sim, time = c(90, 180),
    alpha = c(-0.6, -0.3, 0, 0.3, 0.6),
    intensity.bandwidth = 30,
    knots=c(60,60,60,60,260,460,460,460,460),
    integration.method = 'numerical'
)
})
time.lin <- system.time({
pred.lin <- predict(
    fitted.trt.sim, time = c(90,180),
    alpha = c(-0.6, -0.3, 0, 0.3, 0.6),
    knots=c(60,60,60,60,260,460,460,460,460),
    integration.method = 'linear'
)
})
pred.num$Beta_hat[[2]]
pred.lin$Beta_hat[[3]]
pred.num$term1[[2]] |> head()
pred.lin$term1[[3]] |> head()
pred.num$term2[[2]] |> head()
pred.lin$term2[[3]] |> head()


}
