% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_within_group.R
\name{predict.PCORI_within_group_model}
\alias{predict.PCORI_within_group_model}
\title{Predict mean and variance of the outcome for a PCORI within-group model}
\usage{
\method{predict}{PCORI_within_group_model}(object, time, include.var = TRUE, ..., base = object$base)
}
\arguments{
\item{object}{PCORI_within_group_model object}

\item{time}{Time points of interest}

\item{include.var}{Logical. If TRUE, the variance of the outcome is also returned}
}
\value{
If include.var is TRUE, a tibble with columns time, mean, and var is returned.
otherwise if include.var is FALSE, only the mean vector is returned.
}
\description{
Predict mean and variance of the outcome for a PCORI within-group model
}
