% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensIAT-package.R
\docType{package}
\name{SensIAT-package}
\alias{SensIAT}
\alias{SensIAT-package}
\title{SensIAT: Sensitivity Analysis for Irregular Assessment Times}
\description{
Sensitivity analysis for trials with irregular and informative assessment times, based on a new influence function-based, augmented inverse intensity-weighted estimator.
}
\author{
\strong{Maintainer}: Andrew Redd \email{andrew.redd@hsc.utah.edu} (\href{https://orcid.org/0000-0002-6149-2438}{ORCID})

Other contributors:
\itemize{
  \item Agatha Mallett \email{agatha@geometrian.com} [contributor]
  \item Yujing Gao \email{ygao39@ncsu.edu} (\href{https://orcid.org/0000-0002-6149-2438}{ORCID}) [contributor]
}

}
\keyword{internal}
