% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCORI_example_data.R
\docType{data}
\name{SensIAT_example_data}
\alias{SensIAT_example_data}
\alias{SensIAT_example_fulldata}
\title{SensIAT Example Data}
\format{
A data frame with 779 rows and 4 variables consisting of 200
simulated patients.  Each row in the data represents a visit for the patient.
The columns are:
\describe{
\item{Subject_ID}{A unique identifier for each patient.}
\item{Visit}{The ordinal number of the visit for the patient.  Baseline observation is 0.}
\item{Time}{The time of the visit in days, since baseline.}
\item{Outcome}{The outcome of interest.}
}

A data frame with 779 rows and 4 variables consisting of 200
simulated patients.  Each row in the data represents a visit for the patient.
The columns are:
\describe{
\item{Subject_ID}{A unique identifier for each patient.}
\item{Visit}{The ordinal number of the visit for the patient.  Baseline observation is 0.}
\item{Time}{The time of the visit in days, since baseline.}
\item{Outcome}{The outcome of interest.}
\item{Treatment_group}{Treatment or control group.}
}
}
\usage{
SensIAT_example_data

SensIAT_example_fulldata
}
\description{
A simulated dataset for use in the SensIAT tutorial, testing and documentation.
}
\section{Functions}{
\itemize{
\item \code{SensIAT_example_fulldata}: A simulated dataset with both treatment and control groups.

}}
\keyword{datasets}
