% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensIAT_sim_outcome_modeler_mave.R
\name{SensIAT_sim_outcome_modeler_mave}
\alias{SensIAT_sim_outcome_modeler_mave}
\title{Single Index Model using Mave and Optimizing Bandwidth.}
\usage{
SensIAT_sim_outcome_modeler_mave(
  formula,
  data,
  kernel = "K2_Biweight",
  mave.method = "meanMAVE",
  id = ..id..,
  ...
)
}
\arguments{
\item{formula}{The outcome model formula}

\item{data}{The data to fit the outcome model to.
Should only include follow-up data, i.e. time > 0.}

\item{kernel}{The kernel to use for the outcome model.}

\item{mave.method}{The method to use for the MAVE estimation.}

\item{id}{The patient identifier variable for the data.}

\item{...}{Additional arguments to be passed to \link{optim}.}
}
\value{
Object of class \code{SensIAT::Single-index-outcome-model} which contains the outcome model portion.
}
\description{
Single Index Model using Mave and Optimizing Bandwidth.
}
